+++ decorator/capitalize.rb
@@ -1,0 +1,7 @@
+require './base_decorator'
+
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.strip.capitalize
+  end
+end
