+++ decorator/trimmer.rb
@@ -1,0 +1,7 @@
+require './base_decorator'
+
+class TrimmerDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.strip.slice(0, 10)
+  end
+end
